+++ src/sentry/sentry_apps/api/endpoints/sentry_app_avatar.py
@@ -7,9 +7,7 @@ from sentry.api.base import control_silo_endpoint
 from sentry.api.bases import SentryAppBaseEndpoint
 from sentry.api.bases.avatar import AvatarMixin
 from sentry.models.avatars.sentry_app_avatar import SentryAppAvatar
-from sentry.sentry_apps.api.serializers.parsers.sentry_app_avatar import (
-    RequestSentryAppAvatarSerializer,
-)
+from sentry.sentry_apps.api.parsers.sentry_app_avatar import SentryAppAvatarParser
 from sentry.sentry_apps.api.serializers.sentry_app import SentryAppSerializer
 
 
@@ -22,7 +20,7 @@ class SentryAppAvatarEndpoint(AvatarMixin[SentryAppAvatar], SentryAppBaseEndpoin
     }
     object_type = "sentry_app"
     model = SentryAppAvatar
-    serializer_cls = RequestSentryAppAvatarSerializer
+    serializer_cls = SentryAppAvatarParser
 
     def get(self, request: Request, **kwargs) -> Response:
         return super().get(
