+++ src/sentry/sentry_apps/api/endpoints/sentry_apps.py
@@ -15,7 +15,7 @@ from sentry.api.serializers import serialize
 from sentry.auth.staff import is_active_staff
 from sentry.auth.superuser import is_active_superuser
 from sentry.constants import SentryAppStatus
-from sentry.sentry_apps.api.serializers.parsers.sentry_app import RequestSentryAppSerializer
+from sentry.sentry_apps.api.parsers.sentry_app import SentryAppParser
 from sentry.sentry_apps.api.serializers.sentry_app import (
     SentryAppSerializer as ResponseSentryAppSerializer,
 )
@@ -114,7 +114,7 @@ class SentryAppsEndpoint(SentryAppsBaseEndpoint):
                 status=403,
             )
 
-        serializer = RequestSentryAppSerializer(data=data, access=request.access)
+        serializer = SentryAppParser(data=data, access=request.access)
 
         if serializer.is_valid():
             if data.get("isInternal"):
