+++ src/sentry/sentry_apps/api/parsers/sentry_app_installation.py
@@ -4,7 +4,7 @@ from rest_framework.serializers import Serializer, ValidationError
 from sentry.constants import SentryAppInstallationStatus
 
 
-class RequestSentryAppInstallationSerializer(Serializer):
+class SentryAppInstallationParser(Serializer):
     status = serializers.CharField()
 
     def validate_status(self, new_status):
